Share ImageAsset script utilities
