+++ apps/main/scripts/image-assets/backfill-legacy-images-to-r2.mjs
@@ -1,35 +1,21 @@
-import { PutObjectCommand, S3Client } from "@aws-sdk/client-s3";
-import { PrismaClient } from "@prisma/client";
-import sharp from "sharp";
 import { revalidatePublicCacheForAsset } from "./public-cache-revalidation.mjs";
-
-const DEFAULT_CONCURRENCY = 5;
-const DEFAULT_BATCH_SIZE = 50;
-const DEFAULT_TIMEOUT_SECONDS = 60;
-const DEFAULT_MAX_SOURCE_BYTES = 25 * 1024 * 1024;
-const VARIANT_CACHE_CONTROL = "public, max-age=31536000, immutable";
-const R2_BUCKET_NAME = "daylily-catalog-media";
-const R2_PUBLIC_BASE_URL = "https://media.daylilycatalog.com";
-const SENTRY_DSN =
-  "https://b3773458fec6aa0c594a9c1c73ed046a@o1136137.ingest.us.sentry.io/4508939597643776";
-
-let sentryCaptureException = null;
-if (process.env.NEXT_PUBLIC_SENTRY_ENABLED !== "false") {
-  try {
-    const Sentry = await import("@sentry/nextjs");
-    Sentry.init({ dsn: SENTRY_DSN, enableLogs: true });
-    sentryCaptureException = Sentry.captureException;
-  } catch (error) {
-    console.warn("[image-assets] sentry unavailable", error);
-  }
-}
-
-function readIntEnv(name, fallback) {
-  const value = process.env[name];
-  if (!value) return fallback;
-  const parsed = Number.parseInt(value, 10);
-  return Number.isFinite(parsed) && parsed > 0 ? parsed : fallback;
-}
+import {
+  buildWebpVariants,
+  captureScriptException,
+  createDb,
+  createR2Client,
+  DEFAULT_BATCH_SIZE,
+  DEFAULT_CONCURRENCY,
+  extensionFromUrl,
+  fetchSourceBuffer,
+  getR2BucketName,
+  publicUrlForKey,
+  readIntEnv,
+  runConcurrent,
+  uploadObject,
+  VARIANT_CACHE_CONTROL,
+  variantKeysFromOriginalKey,
+} from "./script-utils.mjs";
 
 function parseArgs() {
   const args = process.argv.slice(2);
@@ -43,126 +29,6 @@ function parseArgs() {
       Number.parseInt(limitValue ?? "", 10) ||
       readIntEnv("IMAGE_ASSET_BACKFILL_BATCH_SIZE", DEFAULT_BATCH_SIZE),
   };
-}
-
-function captureScriptException(error, context) {
-  try {
-    sentryCaptureException?.(error, context);
-  } catch (captureError) {
-    console.warn("[image-assets] sentry capture failed", captureError);
-  }
-}
-
-function requireEnv(name) {
-  const value = process.env[name];
-  if (!value) throw new Error(`${name} is required.`);
-  return value;
-}
-
-async function createDb() {
-  const url = requireEnv("DATABASE_URL");
-
-  if (url.startsWith("file:")) {
-    const { PrismaBetterSqlite3 } = await import(
-      "@prisma/adapter-better-sqlite3"
-    );
-
-    return new PrismaClient({
-      adapter: new PrismaBetterSqlite3(
-        { url },
-        { timestampFormat: "unixepoch-ms" },
-      ),
-      log: ["error", "warn"],
-    });
-  }
-
-  if (url.startsWith("libsql://")) {
-    const { PrismaLibSql } = await import("@prisma/adapter-libsql");
-
-    return new PrismaClient({
-      adapter: new PrismaLibSql(
-        {
-          url,
-          authToken: process.env.TURSO_DATABASE_AUTH_TOKEN,
-        },
-        { timestampFormat: "unixepoch-ms" },
-      ),
-      log: ["error"],
-    });
-  }
-
-  throw new Error(`Unsupported DATABASE_URL: ${url}`);
-}
-
-function createR2Client() {
-  return new S3Client({
-    region: "auto",
-    endpoint: `https://${requireEnv("R2_ACCOUNT_ID")}.r2.cloudflarestorage.com`,
-    credentials: {
-      accessKeyId: requireEnv("R2_ACCESS_KEY_ID"),
-      secretAccessKey: requireEnv("R2_SECRET_ACCESS_KEY"),
-    },
-  });
-}
-
-function publicUrlForKey(key) {
-  const baseUrl = (
-    process.env.R2_PUBLIC_BASE_URL ?? R2_PUBLIC_BASE_URL
-  ).replace(/\/+$/, "");
-  return `${baseUrl}/${key
-    .split("/")
-    .map((segment) => encodeURIComponent(segment))
-    .join("/")}`;
-}
-
-function extensionFromUrl(url) {
-  try {
-    const pathname = new URL(url).pathname;
-    const match = pathname.match(/\.([a-z0-9]+)$/i);
-    return match ? `.${match[1].toLowerCase()}` : ".jpg";
-  } catch {
-    return ".jpg";
-  }
-}
-
-function contentTypeFromExtension(url) {
-  const ext = extensionFromUrl(url);
-  if (ext === ".png") return "image/png";
-  if (ext === ".webp") return "image/webp";
-  if (ext === ".gif") return "image/gif";
-  return "image/jpeg";
-}
-
-function detectImageContentType(buffer, fallback) {
-  if (
-    buffer[0] === 0xff &&
-    buffer[1] === 0xd8 &&
-    buffer[2] === 0xff
-  ) {
-    return "image/jpeg";
-  }
-
-  if (
-    buffer[0] === 0x89 &&
-    buffer[1] === 0x50 &&
-    buffer[2] === 0x4e &&
-    buffer[3] === 0x47
-  ) {
-    return "image/png";
-  }
-
-  if (
-    buffer.subarray(0, 4).toString("ascii") === "RIFF" &&
-    buffer.subarray(8, 12).toString("ascii") === "WEBP"
-  ) {
-    return "image/webp";
-  }
-
-  if (buffer.subarray(0, 3).toString("ascii") === "GIF") {
-    return "image/gif";
-  }
-
-  return fallback;
 }
 
 function originalKeyForImage(image) {
@@ -179,106 +45,6 @@ function originalKeyForImage(image) {
   throw new Error(`Image ${image.id} has no supported owner.`);
 }
 
-function variantKeysFromOriginalKey(originalKey) {
-  const baseKey = originalKey.replace(
-    /\/(?:original|source-original|generated-original)\.[a-z0-9]+$/i,
-    "",
-  );
-  return {
-    displayKey: `${baseKey}/display-800.webp`,
-    thumbKey: `${baseKey}/thumb-200.webp`,
-    blurKey: `${baseKey}/blur-20.webp`,
-  };
-}
-
-async function fetchSourceBuffer(url) {
-  const timeoutSeconds = readIntEnv(
-    "IMAGE_ASSET_BACKFILL_DOWNLOAD_TIMEOUT_SECONDS",
-    DEFAULT_TIMEOUT_SECONDS,
-  );
-  const maxBytes = readIntEnv(
-    "IMAGE_ASSET_BACKFILL_MAX_SOURCE_BYTES",
-    DEFAULT_MAX_SOURCE_BYTES,
-  );
-  const controller = new AbortController();
-  const timer = setTimeout(() => controller.abort(), timeoutSeconds * 1000);
-
-  try {
-    const response = await fetch(url, { signal: controller.signal });
-    if (!response.ok) {
-      throw new Error(`Source download failed: ${response.status}`);
-    }
-
-    const responseContentType = response.headers.get("content-type");
-    const contentLength = Number.parseInt(
-      response.headers.get("content-length") ?? "",
-      10,
-    );
-    if (Number.isFinite(contentLength) && contentLength > maxBytes) {
-      throw new Error(`Source image exceeds ${maxBytes} bytes.`);
-    }
-
-    const buffer = await readResponseBuffer(response, maxBytes);
-
-    return {
-      buffer,
-      contentType: detectImageContentType(
-        buffer,
-        responseContentType?.startsWith("image/")
-          ? responseContentType
-          : contentTypeFromExtension(url),
-      ),
-    };
-  } finally {
-    clearTimeout(timer);
-  }
-}
-
-async function readResponseBuffer(response, maxBytes) {
-  if (!response.body) {
-    const fallbackBuffer = Buffer.from(await response.arrayBuffer());
-    if (fallbackBuffer.byteLength > maxBytes) {
-      throw new Error(`Source image exceeds ${maxBytes} bytes.`);
-    }
-    return fallbackBuffer;
-  }
-
-  const reader = response.body.getReader();
-  const chunks = [];
-  let totalBytes = 0;
-
-  try {
-    for (;;) {
-      const { done, value } = await reader.read();
-      if (done) break;
-      if (!value) continue;
-
-      totalBytes += value.byteLength;
-      if (totalBytes > maxBytes) {
-        throw new Error(`Source image exceeds ${maxBytes} bytes.`);
-      }
-
-      chunks.push(Buffer.from(value));
-    }
-  } finally {
-    reader.releaseLock();
-  }
-
-  return Buffer.concat(chunks, totalBytes);
-}
-
-async function uploadObject(r2, bucket, key, body, contentType, cacheControl) {
-  await r2.send(
-    new PutObjectCommand({
-      Bucket: bucket,
-      Key: key,
-      Body: body,
-      ContentType: contentType,
-      CacheControl: cacheControl,
-    }),
-  );
-}
-
 async function backfillImage({ image, db, r2, bucket, dryRun }) {
   const originalKey = originalKeyForImage(image);
   const variantKeys = variantKeysFromOriginalKey(originalKey);
@@ -292,24 +58,7 @@ async function backfillImage({ image, db, r2, bucket, dryRun }) {
   }
 
   const { buffer, contentType } = await fetchSourceBuffer(image.url);
-  const base = sharp(buffer, { failOn: "error" }).rotate();
-  const [display, thumb, blur] = await Promise.all([
-    base
-      .clone()
-      .resize(800, 800, { fit: "cover" })
-      .webp({ quality: 82 })
-      .toBuffer(),
-    base
-      .clone()
-      .resize(200, 200, { fit: "cover" })
-      .webp({ quality: 78 })
-      .toBuffer(),
-    base
-      .clone()
-      .resize(20, 20, { fit: "cover" })
-      .webp({ quality: 35 })
-      .toBuffer(),
-  ]);
+  const { display, thumb, blur } = await buildWebpVariants(buffer);
 
   await Promise.all([
     uploadObject(r2, bucket, originalKey, buffer, contentType, undefined),
@@ -391,20 +140,6 @@ async function backfillImage({ image, db, r2, bucket, dryRun }) {
   }
 
   console.log("[backfilled]", image.id);
-}
-
-async function runConcurrent(items, concurrency, worker) {
-  let index = 0;
-  const workers = Array.from(
-    { length: Math.min(concurrency, items.length) },
-    async () => {
-      while (index < items.length) {
-        const item = items[index++];
-        await worker(item);
-      }
-    },
-  );
-  await Promise.all(workers);
 }
 
 async function findRetryImages(db, limit) {
@@ -490,7 +225,7 @@ async function main() {
   const args = parseArgs();
   const db = await createDb();
   const r2 = args.dryRun ? null : createR2Client();
-  const bucket = process.env.R2_BUCKET_NAME ?? R2_BUCKET_NAME;
+  const bucket = args.dryRun ? null : getR2BucketName();
   const concurrency = readIntEnv(
     "IMAGE_ASSET_BACKFILL_CONCURRENCY",
     DEFAULT_CONCURRENCY,
